# schemas/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Remote Support Control Plane API
  version: 1.3.0
  description: Implementation contract for attended and managed-host control-plane operations. Media, input, clipboard, chat,
    and file bytes remain peer-to-peer and are not accepted by this API.
servers:
- url: https://api.example.com/v1
  description: Production example; replace by environment configuration
...
security:
- operatorBearer: []
paths:
  /attended-sessions:
    post:
      operationId: createAttendedSession
      summary: Create an attended host session
      tags:
      - Attended
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttendedSessionRequest'
      responses:
        '201':
          description: Pending attended session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendedSessionCreated'
        '400': &id001
          $ref: '#/components/responses/BadRequest'
        '429': &id002
          $ref: '#/components/responses/RateLimited'
        '500': &id003
          $ref: '#/components/responses/InternalError'
  /attended-sessions/resolve:
    post:
      operationId: resolveAttendedSession
      summary: Resolve a support code as an authenticated operator
      tags:
      - Attended
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveAttendedSessionRequest'
      responses:
        '200':
          description: Consent request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '400': *id001
        '401': &id004
          $ref: '#/components/responses/Unauthorized'
        '403': &id006
          $ref: '#/components/responses/Forbidden'
        '409': &id005
          $ref: '#/components/responses/Conflict'
        '429': *id002
        '500': *id003
  /attended-sessions/{sessionId}/pending-consent:
    get:
      operationId: getPendingConsent
      summary: Read the current pending consent request as the host
      tags:
      - Attended
      security:
      - hostBootstrapBearer: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Pending consent request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingConsent'
        '204':
          description: No pending consent request
        '404':
          $ref: '#/components/responses/NotFound'
        '401': *id004
        '409': *id005
        '429': *id002
        '500': *id003
  /attended-sessions/{sessionId}/consent:
    post:
      operationId: decideAttendedConsent
      summary: Approve or reject a pending consent request
      tags:
      - Attended
      security:
      - hostBootstrapBearer: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentDecision'
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          $ref: '#/components/responses/NotFound'
        '400': *id001
        '401': *id004
        '409': *id005
        '429': *id002
        '500': *id003
  /sessions/{sessionId}:
    get:
      operationId: getSession
      summary: Read authorized session metadata
      tags:
      - Sessions
      security:
      - dpopProof: []
        peerDpopAccessToken: []
      - operatorBearer: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          $ref: '#/components/responses/NotFound'
        '401': *id004
        '403': *id006
        '500': *id003
    delete:
      operationId: endSession
      summary: End a session
      tags:
      - Sessions
      security:
      - dpopProof: []
        peerDpopAccessToken: []
      - operatorBearer: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: Ended or already ended
        '404':
          $ref: '#/components/responses/NotFound'
        '401': *id004
        '403': *id006
        '409': *id005
        '500': *id003
  /sessions/{sessionId}/peer-authorization:
    post:
      operationId: exchangePeerAuthorization
      summary: Exchange a bootstrap credential and key proof for a scoped peer token
      tags:
      - Sessions
      security:
      - sessionBootstrapBearer: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeerAuthorizationRequest'
      responses:
        '200':
          description: Peer authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerAuthorization'
        '404':
          $ref: '#/components/responses/NotFound'
        '400': *id001
        '401': *id004
        '403': *id006
        '409': *id005
        '429': *id002
        '500': *id003
  /sessions/{sessionId}/signaling-tickets:
    post:
      operationId: createSignalingTicket
      summary: Create a one-time signaling WebSocket ticket
      tags:
      - Sessions
      security:
      - dpopProof: []
        peerDpopAccessToken: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Signaling ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalingTicket'
        '404':
          $ref: '#/components/responses/NotFound'
        '401': *id004
        '403': *id006
        '409': *id005
        '429': *id002
        '500': *id003
  /sessions/{sessionId}/turn-credentials:
    post:
      operationId: createTurnCredentials
      summary: Issue short-lived TURN credentials for the current peer
      tags:
      - Sessions
      security:
      - dpopProof: []
        peerDpopAccessToken: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: TURN credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnCredentials'
        '404':
          $ref: '#/components/responses/NotFound'
        '401': *id004
        '403': *id006
        '409': *id005
        '429': *id002
        '500': *id003
...
  /sessions/{sessionId}/scope-revocations:
    post:
      operationId: revokeSessionScopes
      summary: Revoke one or more granted scopes
      tags:
      - Sessions
      security:
      - dpopProof: []
        peerDpopAccessToken: []
      - operatorBearer: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScopeRevocationRequest'
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          $ref: '#/components/responses/NotFound'
        '400': *id001
        '401': *id004
        '403': *id006
        '409': *id005
        '500': *id003
  /sessions/{sessionId}/termination:
    post:
      operationId: terminateSession
      summary: Immediately terminate an authorized session as either peer
      tags:
      - Sessions
      security:
      - dpopProof: []
        peerDpopAccessToken: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionTerminationRequest'
      responses:
        '200':
          description: Terminated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400': *id001
        '401': *id004
        '403': *id006
        '404':
          $ref: '#/components/responses/NotFound'
        '409': *id005
        '429': *id002
        '500': *id003
...
  /sessions/{sessionId}/peer-authorization-challenges:
    post:
      operationId: createPeerAuthorizationChallenge
      summary: Create a single-use proof challenge for the bootstrap-bound peer
      tags:
      - Sessions
      security:
      - sessionBootstrapBearer: []
      parameters:
      - $ref: '#/components/parameters/SessionId'
      responses:
        '201':
          description: Peer authorization challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerAuthorizationChallenge'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    operatorBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token for an authenticated operator/admin.
    hostBootstrapBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Short-lived attended-host bootstrap credential.
    sessionBootstrapBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Host or operator bootstrap credential used only for peer authorization exchange.
    dpopProof:
      type: apiKey
      in: header
      name: DPoP
      description: RFC 9449 proof JWT. Required together with peer/device access tokens; htm, htu, iat, jti, ath and server
        nonce are validated and replay-cached.
    peerDpopAccessToken:
      type: http
      scheme: DPoP
      description: Short-lived session/peer/epoch/scope-bound access token containing a cnf.jkt key thumbprint. Every use
        requires a matching RFC 9449 DPoP proof.
    deviceDpopAccessToken:
      type: http
      scheme: DPoP
      description: Short-lived installed-device access token containing device/key/authorization-version claims and cnf.jkt.
        Every use requires a matching RFC 9449 DPoP proof.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: ^"[0-9]+"$
      description: Quoted state or resource version.
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    RouteClass:
      type:
      - string
      - 'null'
      enum:
      - DIRECT
      - TURN_UDP
      - TURN_TCP
      - TURN_TLS
      - null
...
    CreateAttendedSessionRequest:
      type: object
      additionalProperties: false
      required:
      - hostEphemeralPublicKey
      - clientVersion
      - capabilities
      properties:
        hostEphemeralPublicKey:
          $ref: '#/components/schemas/PublicKeyJwk'
        clientVersion:
          type: string
          minLength: 1
          maxLength: 64
        installationInstanceId:
          type: string
          format: uuid
        locale:
          type: string
          maxLength: 32
        capabilities:
          $ref: '#/components/schemas/ClientCapabilities'
    AttendedSessionCreated:
      type: object
      additionalProperties: false
      required:
      - sessionId
      - supportCode
      - expiresAt
      - hostBootstrapToken
      - hostPeerId
      - state
      - stateVersion
      - pendingEventsUrl
      properties:
        sessionId:
          type: string
          format: uuid
        supportCode:
          $ref: '#/components/schemas/SupportCode'
        expiresAt:
          type: string
          format: date-time
        hostBootstrapToken:
          type: string
          minLength: 32
        hostPeerId:
          type: string
          format: uuid
        state:
          $ref: '#/components/schemas/SessionState'
        stateVersion:
          type: integer
          minimum: 1
        pendingEventsUrl:
          type: string
          format: uri
    ResolveAttendedSessionRequest:
      type: object
      additionalProperties: false
      required:
      - supportCode
      - requestedScopes
      - operatorEphemeralPublicKey
      - clientVersion
      properties:
        supportCode:
          $ref: '#/components/schemas/SupportCode'
        requestedScopes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Scope'
        operatorEphemeralPublicKey:
          $ref: '#/components/schemas/PublicKeyJwk'
        clientVersion:
          type: string
          minLength: 1
          maxLength: 64
        capabilities:
          $ref: '#/components/schemas/ClientCapabilities'
    ConsentRequest:
      type: object
      additionalProperties: false
      required:
      - sessionId
      - consentRequestId
      - operator
      - requestedScopes
      - expiresAt
      - stateVersion
      - operatorBootstrapToken
      properties:
        sessionId:
          type: string
          format: uuid
        consentRequestId:
          type: string
          format: uuid
        operator:
          type: object
          additionalProperties: false
          required:
          - displayName
          - tenantDisplayName
          properties:
            displayName:
              type: string
              minLength: 1
              maxLength: 200
            tenantDisplayName:
              type: string
              minLength: 1
              maxLength: 200
            verifiedTenant:
              type: boolean
        requestedScopes:
          type: array
          items:
            $ref: '#/components/schemas/Scope'
        expiresAt:
          type: string
          format: date-time
        stateVersion:
          type: integer
          minimum: 1
        operatorBootstrapToken:
          type: string
          minLength: 32
          description: Returned only to the authenticated operator caller.
    PendingConsent:
      type: object
      additionalProperties: false
      required:
      - sessionId
      - consentRequestId
      - operator
      - requestedScopes
      - expiresAt
      - stateVersion
      - consentNonce
      - hostEphemeralKeyThumbprint
      properties:
        sessionId:
          type: string
          format: uuid
        consentRequestId:
          type: string
          format: uuid
        operator:
          type: object
          additionalProperties: false
          required:
          - displayName
          - tenantDisplayName
          - verifiedTenant
          properties:
            displayName:
              type: string
            tenantDisplayName:
              type: string
            verifiedTenant:
              type: boolean
        requestedScopes:
          type: array
          items:
            $ref: '#/components/schemas/Scope'
        expiresAt:
          type: string
          format: date-time
        stateVersion:
          type: integer
        consentNonce:
          type: string
          minLength: 16
          maxLength: 256
        hostEphemeralKeyThumbprint:
          type: string
          pattern: ^[A-Za-z0-9_-]{43}$
          description: RFC 7638 SHA-256 JWK thumbprint in unpadded base64url.
    ConsentDecision:
      type: object
      additionalProperties: false
      required:
      - consentRequestId
      - approved
      - grantedScopes
      - consentNonce
      - decisionProof
      properties:
        consentRequestId:
          type: string
          format: uuid
        approved:
          type: boolean
        grantedScopes:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Scope'
        consentNonce:
          type: string
          minLength: 16
          maxLength: 256
        decisionProof:
          $ref: '#/components/schemas/DetachedProof'
    Session:
      type: object
      additionalProperties: false
      required:
      - id
      - sessionType
      - state
      - stateVersion
      - permissionRevision
      - requestedScopes
      - grantedScopes
      - createdAt
      - expiresAt
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type:
          - string
          - 'null'
          format: uuid
        sessionType:
          type: string
          enum:
          - ATTENDED
          - MANAGED_ATTENDED
          - UNATTENDED
        state:
          $ref: '#/components/schemas/SessionState'
        stateVersion:
          type: integer
          minimum: 1
        permissionRevision:
          type: integer
          minimum: 0
        transportEpoch:
          type: integer
          minimum: 0
        requestedScopes:
          type: array
          items:
            $ref: '#/components/schemas/Scope'
        grantedScopes:
          type: array
          items:
            $ref: '#/components/schemas/Scope'
        createdAt:
          type: string
          format: date-time
        authorizedAt:
          type:
          - string
          - 'null'
          format: date-time
        connectedAt:
          type:
          - string
          - 'null'
          format: date-time
        endedAt:
          type:
          - string
          - 'null'
          format: date-time
        expiresAt:
          type: string
          format: date-time
        endReason:
          type:
          - string
          - 'null'
          maxLength: 128
        routeClass:
          $ref: '#/components/schemas/RouteClass'
...
    PeerAuthorizationRequest:
      type: object
      additionalProperties: false
      required:
      - challengeId
      - role
      - proof
      properties:
        role:
          $ref: '#/components/schemas/PeerRole'
        proof:
          $ref: '#/components/schemas/ProofOfPossession'
        challengeId:
          type: string
          format: uuid
    PeerAuthorization:
      type: object
      additionalProperties: false
      required:
      - sessionId
      - peerId
      - role
      - peerToken
      - grantedScopes
      - permissionRevision
      - transportEpoch
      - expiresAt
      - remotePeerId
      - remoteRole
      - remoteEphemeralPublicKey
      - remoteKeyThumbprint
      - authorizationContextSha256
      properties:
        sessionId:
          type: string
          format: uuid
        peerId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/PeerRole'
        peerToken:
          type: string
          minLength: 32
        grantedScopes:
          type: array
          items:
            $ref: '#/components/schemas/Scope'
        permissionRevision:
          type: integer
          minimum: 0
        transportEpoch:
          type: integer
          minimum: 0
        expiresAt:
          type: string
          format: date-time
        remotePeerId:
          type: string
          format: uuid
        remoteRole:
          $ref: '#/components/schemas/PeerRole'
        remoteEphemeralPublicKey:
          $ref: '#/components/schemas/PublicKeyJwk'
        remoteKeyThumbprint:
          type: string
          minLength: 40
          maxLength: 128
        authorizationContextSha256:
          type: string
          minLength: 43
          maxLength: 43
    SignalingTicket:
      type: object
      additionalProperties: false
      required:
      - ticket
      - signalingUrl
      - expiresAt
      - transportEpoch
      properties:
        ticket:
          type: string
          minLength: 32
        signalingUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
        transportEpoch:
          type: integer
          minimum: 0
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authorization or policy denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: State or version conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: Rate or abuse limit reached
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'